requirements.txt
pandas
openpyxl
requests
python-dotenv
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
groq